edit wbs_DB.py; edit delete_all_wbs func

+++ wbs_DB.py
@@ -52,6 +52,13 @@ def delete_all_wbs(pid):
     cur = connection.cursor(pymysql.cursors.DictCursor)
 
     try:
+        cur.execute("SELECT COUNT(*) AS cnt FROM progress WHERE p_no = %s", (pid,))
+        result = cur.fetchone()
+
+        if result['cnt'] == 0:
+            print(f"Error [delete_all_wbs] : Progress data does not exist in Project UID {pid}.")
+            return False
+        
         cur.execute("DELETE FROM progress WHERE p_no = %s", (pid,))
         connection.commit()
         return True
